+++ backend/src/main/java/com/icehufs/icebreaker/domain/codingzone/dto/request/GroupInfUpdateRequestDto.java
@@ -34,5 +34,5 @@ public class GroupInfUpdateRequestDto {
     private String weekDay; //요일
 
     @NotNull
-    private Integer grade;
+    private Integer subjectId;
 }
