+++ backend/src/main/java/com/icehufs/icebreaker/domain/codingzone/repository/CodingZoneRegisterRepository.java
@@ -10,8 +10,8 @@ import java.util.List;
 public interface CodingZoneRegisterRepository extends JpaRepository<CodingZoneRegister, Integer> {
     CodingZoneRegister findByClassNumAndUserEmail(Integer classNum, String userEmail);
     CodingZoneRegister findByRegistrationId(Integer registrationId);
-    List<CodingZoneRegister> findByGrade(int grade);
+    List<CodingZoneRegister> findBySubjectId(int subjectId);
     List<CodingZoneRegister> findByUserEmail(String userEmail);
     List<CodingZoneRegister> findAllByOrderByUserStudentNumAsc();
-    List<CodingZoneRegister> findByGradeOrderByUserStudentNumAsc(int grade);
+    List<CodingZoneRegister> findBySubjectIdOrderByUserStudentNumAsc(int subjectId);
 }
